Properly sign the access_token request with the request secret
Twitter does not currently require this, it seems to ignore the
signature on access_token requests alltogether (which is in violation
with the OAuth spec and their own documentation). To prevent making this
code a bad example and problems when Twitter ever becomes compliant,
we'll do this the proper way.